requirements.txt
apturl==0.5.2
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
certifi==2018.11.29
chardet==3.0.4
checkbox-support==0.22
command-not-found==0.3
cryptography==1.2.3
defer==1.0.6
defusedxml==0.5.0
dell-recovery==0.0.0
Django==2.1.4
django-allauth==0.38.0
django-filter==2.0.0
django-recaptcha2==1.3.1
django-recaptcha3==0.0.1
djangorestframework==3.9.0
feedparser==5.1.3
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.8
Jinja2==2.8
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==3.0.1
MarkupSafe==0.23
oauthlib==2.1.0
onboard==1.2.0
padme==1.1.1
pexpect==4.0.1
Pillow==5.3.0
plainbox==0.25
progressbar==2.3
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.0.3
python-apt==1.1.0b1+ubuntu0.16.4.2
python-debian==0.1.27
python-systemd==231
python3-openid==3.1.0
pytz==2018.7
pyxdg==0.25
reportlab==3.3.0
requests==2.21.0
requests-oauthlib==1.0.0
sessioninstaller==0.0.0
six==1.10.0
stripe==2.16.0
system-service==0.3
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
unity-tweak-tool==0.0.7
urllib3==1.24.1
usb-creator==0.3.0
virtualenv==16.1.0
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3